requirements.txt
langchain-google-genai>=0.0.5
python-dotenv>=0.19.0
duckduckgo-search>=4.1.1
newsapi-python>=0.2.7
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.1
requests>=2.31.0